Add an HTML reporter based on the element hierarchy.
The default HTML reporter of Jnario creates HTML pages with a specific
style associated to each type of specification element (fact,
description, etc).

This commit adds an extension of this reporter, which is generating the
HTML pages by using only the depth of the elements in the specification
(not the type of the element).

Additionnally, this new reporter replaces any occurrence of "@outline"
or "<!-- OUTPUT OUTLINE -->" by a computed outline from the top level
elements in the specification.

The new reporter replaces several HTML tags in the specification by the
correct HTML code in the generated pages. For example,
"<note>text<note>" permits to put a note paragraph with the appropriate
icon.